+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_transport_serializer_config_invalid.php
@@ -1,0 +1,14 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'messenger' => [
+        'transports' => [
+            'foo' => [
+                'dsn' => 'null://',
+                'symfony_serializer' => [
+                    'format' => 'xml',
+                ]
+            ],
+        ],
+    ],
+]);